fix build on 64 bit hosts
llvm-svn: 25591

+++ llvm/lib/CodeGen/DwarfWriter.cpp
@@ -645,7 +645,7 @@ void DIEAbbrev::Emit(const DwarfWriter &DW) const {
 #ifndef NDEBUG
   void DIEAbbrev::print(std::ostream &O) {
     O << "Abbreviation @"
-      << std::hex << (unsigned)this << std::dec
+      << std::hex << (unsigned long)this << std::dec
       << "  "
       << TagString(Tag)
       << " "
